Read a format capture as the reference it is
The liveness gate counted identifiers from code and dropped every string
literal, so an item named only through an inline format capture read as
referenced by tests alone. `format!("{NAME}")` resolves `NAME` in the
surrounding scope and deleting the item breaks the build, so the capture is
counted and an escaped brace is not.

The SAFETY comment on the CUDA barrier reset sat above the probe call rather
than the unsafe closure it justifies. The visitor arm macro suppressed
`unused_variables` at the item; the binding is read instead, which an
`expect` cannot do because most arms do use the receiver.

Twelve driver and fixture expects state the corrective action, and twenty
clippy findings across nine crates are resolved. Two asserts over a const
array folded to a constant and proved nothing; both now count what their
loop did.

Six visual operations recorded `uncharacterized` while the reference oracle
refutes every law family their shape admits. They record no-legal-rewrite,
which is what the run derives.

Property-test coverage records the measured floor. The deleted fused
frontier clear property is restored in the crate that now owns the builder,
and the padded shared extent in the PTX prologue gains a property over the
element types read from the scalar roster at run time.

+++ vyre-foundation/src/serial/wire/encode/to_wire/mod.rs
@@ -229,8 +229,8 @@ fn reject_non_roundtrippable_shapes(
                     WireEncodeErr::Dynamic(Box::new(buf))
                 })?
             };
-            let start = range.start as u64;
-            let end = range.end as u64;
+            let start = range.start;
+            let end = range.end;
             if start > end {
                 let mut buf = arrayvec::ArrayString::<256>::new();
                 let mut tmp = itoa::Buffer::new();
